fix warning errors

+++ src/cpp/illum_cli.cpp
@@ -37,7 +37,7 @@ MPI_Info info = MPI_INFO_NULL;
 template <typename T>
 std::pair<arma::uvec, arma::uvec> get_rowinds_and_colptrs(arma::SpMat<T> const & S) {
   arma::uvec rowinds(S.n_nonzero), colptrs(S.n_cols + 1);
-  int i = 0, j = 0;
+  size_t i = 0, j = 0;
   colptrs(0) = 0;
   for (auto it = S.begin(); it != S.end(); ++it) {
     rowinds(i++) = it.row();
@@ -64,8 +64,8 @@ void write_csc_inds(arma::SpMat<T> const & S, const char * path) {
 void write_coo(arma::sp_umat const & S, const char * path) {
   std::ofstream f;
   f.open(path);
-  for (int j = 0; j < S.n_cols; ++j) {
-    for (int i = 0; i < S.n_rows; ++i) {
+  for (size_t j = 0; j < S.n_cols; ++j) {
+    for (size_t i = 0; i < S.n_rows; ++i) {
       if (S(i, j)) {
         f << i << ", " << j << ", " << S(i, j) << std::endl;
       }
